Move ProjectionPushDown util methods to plugin toolkit
