+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveApplyProjectionUtil.java
@@ -13,109 +13,18 @@
  */
 package io.trino.plugin.hive;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import io.trino.spi.connector.ColumnHandle;
-import io.trino.spi.expression.Call;
-import io.trino.spi.expression.ConnectorExpression;
-import io.trino.spi.expression.Constant;
-import io.trino.spi.expression.FieldDereference;
-import io.trino.spi.expression.Variable;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 
-import static com.google.common.collect.ImmutableList.toImmutableList;
-import static java.util.Objects.requireNonNull;
+import static io.trino.plugin.base.projection.ApplyProjectionUtil.ProjectedColumnRepresentation;
 
 public final class HiveApplyProjectionUtil
 {
     private HiveApplyProjectionUtil() {}
-
-    public static List<ConnectorExpression> extractSupportedProjectedColumns(ConnectorExpression expression)
-    {
-        requireNonNull(expression, "expression is null");
-        ImmutableList.Builder<ConnectorExpression> supportedSubExpressions = ImmutableList.builder();
-        fillSupportedProjectedColumns(expression, supportedSubExpressions);
-        return supportedSubExpressions.build();
-    }
-
-    private static void fillSupportedProjectedColumns(ConnectorExpression expression, ImmutableList.Builder<ConnectorExpression> supportedSubExpressions)
-    {
-        if (isPushDownSupported(expression)) {
-            supportedSubExpressions.add(expression);
-            return;
-        }
-
-        // If the whole expression is not supported, look for a partially supported projection
-        for (ConnectorExpression child : expression.getChildren()) {
-            fillSupportedProjectedColumns(child, supportedSubExpressions);
-        }
-    }
-
-    @VisibleForTesting
-    static boolean isPushDownSupported(ConnectorExpression expression)
-    {
-        return expression instanceof Variable ||
-                (expression instanceof FieldDereference fieldDereference && isPushDownSupported(fieldDereference.getTarget()));
-    }
-
-    public static ProjectedColumnRepresentation createProjectedColumnRepresentation(ConnectorExpression expression)
-    {
-        ImmutableList.Builder<Integer> ordinals = ImmutableList.builder();
-
-        Variable target;
-        while (true) {
-            if (expression instanceof Variable variable) {
-                target = variable;
-                break;
-            }
-            if (expression instanceof FieldDereference dereference) {
-                ordinals.add(dereference.getField());
-                expression = dereference.getTarget();
-            }
-            else {
-                throw new IllegalArgumentException("expression is not a valid dereference chain");
-            }
-        }
-
-        return new ProjectedColumnRepresentation(target, ordinals.build().reverse());
-    }
-
-    /**
-     * Replace all connector expressions with variables as given by {@param expressionToVariableMappings} in a top down manner.
-     * i.e. if the replacement occurs for the parent, the children will not be visited.
-     */
-    public static ConnectorExpression replaceWithNewVariables(ConnectorExpression expression, Map<ConnectorExpression, Variable> expressionToVariableMappings)
-    {
-        if (expressionToVariableMappings.containsKey(expression)) {
-            return expressionToVariableMappings.get(expression);
-        }
-
-        if (expression instanceof Constant || expression instanceof Variable) {
-            return expression;
-        }
-
-        if (expression instanceof FieldDereference fieldDereference) {
-            ConnectorExpression newTarget = replaceWithNewVariables(fieldDereference.getTarget(), expressionToVariableMappings);
-            return new FieldDereference(expression.getType(), newTarget, fieldDereference.getField());
-        }
-
-        if (expression instanceof Call call) {
-            return new Call(
-                    call.getType(),
-                    call.getFunctionName(),
-                    call.getArguments().stream()
-                            .map(argument -> replaceWithNewVariables(argument, expressionToVariableMappings))
-                            .collect(toImmutableList()));
-        }
-
-        // We cannot skip processing for unsupported expression shapes. This may lead to variables being left in ProjectionApplicationResult
-        // which are no longer bound.
-        throw new UnsupportedOperationException("Unsupported expression: " + expression);
-    }
 
     /**
      * Returns the assignment key corresponding to the column represented by {@param projectedColumn} in the {@param assignments}, if one exists.
@@ -155,52 +64,5 @@ public final class HiveApplyProjectionUtil
         }
 
         return Optional.empty();
-    }
-
-    public static class ProjectedColumnRepresentation
-    {
-        private final Variable variable;
-        private final List<Integer> dereferenceIndices;
-
-        public ProjectedColumnRepresentation(Variable variable, List<Integer> dereferenceIndices)
-        {
-            this.variable = requireNonNull(variable, "variable is null");
-            this.dereferenceIndices = ImmutableList.copyOf(requireNonNull(dereferenceIndices, "dereferenceIndices is null"));
-        }
-
-        public Variable getVariable()
-        {
-            return variable;
-        }
-
-        public List<Integer> getDereferenceIndices()
-        {
-            return dereferenceIndices;
-        }
-
-        public boolean isVariable()
-        {
-            return dereferenceIndices.isEmpty();
-        }
-
-        @Override
-        public boolean equals(Object obj)
-        {
-            if (this == obj) {
-                return true;
-            }
-            if ((obj == null) || (getClass() != obj.getClass())) {
-                return false;
-            }
-            ProjectedColumnRepresentation that = (ProjectedColumnRepresentation) obj;
-            return Objects.equals(variable, that.variable) &&
-                    Objects.equals(dereferenceIndices, that.dereferenceIndices);
-        }
-
-        @Override
-        public int hashCode()
-        {
-            return Objects.hash(variable, dereferenceIndices);
-        }
     }
 }
